+++ src/app/UI/widgets/button/button.component.ts
@@ -6,11 +6,17 @@ import { Component, EventEmitter, Input, Output } from '@angular/core';
   styleUrls: ['./button.component.scss']
 })
 export class ButtonComponent {
+  button = {
+    primary: 'primary',
+    secondary: 'secondary',
+    smallPrimary: 'small-primary',
+    smallSecondary: 'small-secondary',
+  }
   @Input() buttonTittle: string = '';
-  @Input() primaryButton: boolean = true;
+  @Input() buttonType: string = this.button.primary;
   @Output() onClick: EventEmitter<string> = new EventEmitter<string>();
 
-  buttonClicked() {
+  buttonClicked(): void {
     this.onClick.emit('button clicked');
   }
 }